+++ CHANGELOG.md
@@ -16,6 +16,12 @@ Semantic versioning in our case means:
   But, in the future we might change the configuration names/logic,
   change the client facing API, change code conventions significantly, etc.
 
+## WIP
+
+### Bugfixes
+
+- Fixes false positive `WPS457` for ``while True`` loop with ``await`` expressions, #3753
+
 
 ## 1.5.0