+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Semantic versioning in our case means:
 
 - Fixes false positive `WPS457` for ``while True`` loop with ``await`` expressions, #3753
 - Fixes the false positive `WPS617` by assigning a function that receives a lambda expression as a parameter.
+- Fixes false positive `WPS430` for whitelisted nested functions, #3589
 
 
 ## 1.5.0