+++ CHANGELOG.md
@@ -17,12 +17,14 @@ Semantic versioning in our case means:
   change the client facing API, change code conventions significantly, etc.
 
 
-## WIP
+## 1.6.1
 
 ### Bugfixes
 
-- Fixes false positive `WPS366` allowing the use of a single constant in `or`, #3610
+- Fixes false positive `WPS366` allowing the use
+  of a single constant in `or`, #3610
 - Fixes the false positive `WPS330` when alternating unary operators, #3594
+
 
 ## 1.6.0